Move around some of the types
