+++ packages/data-mate/src/column/transforms/toUpperCase.ts
@@ -1,5 +1,5 @@
-import { VectorType } from '../../vector';
-import { ColumnFnMode, ColumnTransformConfig } from '../interfaces';
+import { VectorIteratorMode, VectorType } from '../../vector';
+import { ColumnTransformConfig } from '../interfaces';
 
 /**
  * Converts strings to upper case
@@ -12,7 +12,7 @@ import { ColumnFnMode, ColumnTransformConfig } from '../interfaces';
 export const toUpperCaseConfig: ColumnTransformConfig<string> = {
     create() {
         return {
-            mode: ColumnFnMode.EACH_VALUE,
+            mode: VectorIteratorMode.EACH_VALUE,
             skipNulls: true,
             fn(input) {
                 return input.toUpperCase();
