+++ packages/data-mate/src/vector/interfaces.ts
@@ -1,0 +1,63 @@
+import { Maybe } from '@terascope/types';
+
+/**
+ * The Vector Type, this will change how the data is stored and read
+*/
+export enum VectorType {
+    /**
+     * Currently this operates like String
+     * but I imagine will be expanding it.
+     * But will need to add format options
+    */
+    Date = 'Date',
+    String = 'String',
+    Int = 'Int',
+    Float = 'Float',
+    BigInt = 'BigInt',
+    Boolean = 'Boolean',
+    GeoPoint = 'GeoPoint',
+    GeoJSON = 'GeoJSON',
+    /** @todo */
+    Object = 'Object',
+    /**
+     * Arbitrary data can be stored with this
+    */
+    Any = 'Any',
+    /**
+     * The list type is used for fields marked as Arrays
+     * where each item in the Vector is a child element
+    */
+    List = 'List',
+}
+
+/**
+ * A data type agnostic in-memory representation of the data
+ * for a Vector and potential indices.
+ * This should be generated by the builder.
+*/
+export type Data<T> = Readonly<{
+    readonly values: readonly Maybe<T>[];
+}>
+
+/**
+ * A mode to describe which value to pass when iterating over a Vector
+*/
+export enum VectorIteratorMode {
+    /**
+     * Just the values from Vector, automatically deals with ListVectors and skips nulls.
+     * This is probably the most common.
+     *
+     * @todo This will be called with only unique values.
+    */
+    EACH_VALUE,
+
+    /**
+     * A values from the Vector
+    */
+    EACH,
+
+    /**
+     * The entire Vector be passed
+    */
+    ALL,
+}
