+++ packages/data-mate/src/vector/types/object.ts
@@ -1,4 +1,5 @@
-import { Vector, VectorOptions, VectorType } from '../vector';
+import { Vector, VectorOptions } from '../vector';
+import { VectorType } from '../interfaces';
 
 /**
  * @todo we need an to serialize to JSON correctly
