+++ packages/data-mate/src/vector/types/string.ts
@@ -1,4 +1,5 @@
-import { Vector, VectorOptions, VectorType } from '../vector';
+import { Vector, VectorOptions } from '../vector';
+import { VectorType } from '../interfaces';
 
 export class StringVector extends Vector<string> {
     constructor(options: VectorOptions<string>) {
