+++ packages/data-mate/src/vector/vector.ts
@@ -1,6 +1,7 @@
 import {
     DataTypeFieldConfig, DataTypeFields, FieldType, Maybe, Nil, SortOrder
 } from '@terascope/types';
+import { Data, VectorType } from './interfaces';
 
 /**
  * An immutable typed Array class with a constrained API.
@@ -181,36 +182,6 @@ export function isVector<T>(input: unknown): input is Vector<T> {
 }
 
 /**
- * The Vector Type, this will change how the data is stored and read
-*/
-export enum VectorType {
-    /**
-     * Currently this operates like String
-     * but I imagine will be expanding it.
-     * But will need to add format options
-    */
-    Date = 'Date',
-    String = 'String',
-    Int = 'Int',
-    Float = 'Float',
-    BigInt = 'BigInt',
-    Boolean = 'Boolean',
-    GeoPoint = 'GeoPoint',
-    GeoJSON = 'GeoJSON',
-    /** @todo */
-    Object = 'Object',
-    /**
-     * Arbitrary data can be stored with this
-    */
-    Any = 'Any',
-    /**
-     * The list type is used for fields marked as Arrays
-     * where each item in the Vector is a child element
-    */
-    List = 'List',
-}
-
-/**
  * Serialize a value to a JSON compatible format (so it can be JSON stringified)
 */
 export type ValueToJSONFn<T> = (value: T, thisArg?: Vector<T>) => any;
@@ -227,12 +198,3 @@ export interface VectorOptions<T> {
 
 export type JSONValue<T> = T extends Vector<infer U> ? U[] : T;
 export type MaybeJSONValue<T> = Maybe<T>|Maybe<JSONValue<T>>;
-
-/**
- * A data type agnostic in-memory representation of the data
- * for a Vector and potential indices.
- * This should be generated by the builder.
-*/
-export type Data<T> = Readonly<{
-    readonly values: readonly Maybe<T>[];
-}>
